Fix bugs in handling of Date objects
The attempt to always return POSIXct in ggplot.ts() and try_tibble() caused time shifts for dates. We now return Dates, when this is the class of the xts index or when it is "yearqtr". Possibly "yearmon" should be also converted into Date. Documentation and examples revised. stat_peaks() and stat_valleys() were mishandling Date mapped onto x. Examples added to annotate() docs. Test cases for annotate() implemented.