Fix handling of too few distinct x values
Check as previously implemented would return an empty tibble, which resulted in no output generated for other groups. Now we override formula and return the fit to y ~ 1, i.e., the mean.
Rounding of rr and rr.adj is changed so that if their value is exactly zero they are reported as such.