Revise stat_quant_eq() and its test cases
The positioning of the equations for multiple quantiles was working only for character arguments for label.x and label.y.  Added code to handle numeric values, either as data coordinates or npc. Added also example for angle. != 0. Numeric values  in npc are now used as is, with no addition of 5% margin as earlier. This was also changes for stat_oly_eq().