Quoted strings for numbers in expressions
In this way numbers are parsed into strings, avoiding dropping of trailing zeros and allowing formatting with bold(), etc. Done for all  relevant labels in stat_poly_eq()  and stat_quant_eq() except equation coefs. I need to consider if this should be the default or not and how to do the same with equations.