Revise stat_poly_eq() to return the coef estimates
For now the coef.ls list continues to be returned as it contains the stats in addition to estimates for each coef. On the other hand the list is not handled correctly by 'ggplot2' when the plots contain facets. The coef columns are named b_0, b_1, etc. b_0 is always returned, even if forced to zero.