Compatibility with 'ggplot2' (>= 3.5.1.9000).
'ggplot2' now accepts lambdas as scale name. As a call was being passed instead of an expression, when not passing a plain character string, errors were triggered. I enclosed in the definition of FC_name() the call to bquote() in a call to as.expression() and the problem seems to be solved.
Update `scale_x_logFC()`and `scale_y_logFC()` for compatibility with 'ggplot2' (>= 3.5.1.9000).