find_peaks update unit tests and examples in docs
The logic and expected values for parameters are somehow different to those in the previous, never released approach. This made necessary to update a few unit tests. In addition the units tests revealed that sanity checks on threshold arguments were incomplete. These are now more thorough and match the valid ranges in the new implementation.