LaTeX support with 'xdvir'
Factored out and polished the checks for 'output.type' arguments. Added 'latex.eqn' and 'latex.deqn'. Added and updated unit tests for LaTeX labels. Fixed several bugs, a couple in the LaTeX formatting of labels and a couple in the unit tests. Improved the documentation for output.type, using inheritance.
Updated snaps of unit tests for upcoming 'ggpp' 0.6.0.
Started a new on-line only article demonstrating the use of 'ggpmisc' with 'xdvir'. (lots of progress in a short time.)